build only 32-bit or only 64-bit for travis

+++ build-setup-travis.sh
@@ -11,17 +11,33 @@ case $(arch) in
 esac
 
 set -x
+
+build64=true
+build32=true
+case $1 in
+    ONLYARCH=x86_64) build32=false;;
+    ONLYARCH=i386)   build64=false;;
+esac
+
+if $build32; then
+    echo Building 32-bit.
+fi
+if $build64; then
+    echo Building 64-bit.
+fi
+
+
 rm -f ${BUILDTOP}/host
 ln -f -s ${BUILDTOP}/${HOST} ${BUILDTOP}/host
 mkdir -p ${BUILDTOP}/${HOST}
 
     if [ ! -d $BUILDTOP/mbedtls ]; then (cd ${BUILDTOP} && git clone -b mcr_add_otherName https://github.com/mcr/mbedtls.git ); fi
 
-    if [ ! -d "${BUILDTOP}/x86_64/mbedtls" ]; then
+    if $build64 && [ ! -d "${BUILDTOP}/x86_64/mbedtls" ]; then
         (cd ${BUILDTOP} && rm -rf x86_64/mbedtls && mkdir -p x86_64/mbedtls && cd x86_64/mbedtls && cmake -DCMAKE_C_FLAGS:STRING=-m64 -DCMAKE_INSTALL_PREFIX=$BUILDTOP/x86_64 ../../mbedtls && make CFLAGS='-m64 --coverage -g3 -O0' && make install)
     fi
 
-    if [ ! -d "${BUILDTOP}/i386/mbedtls" ]; then
+    if $build32 && [ ! -d "${BUILDTOP}/i386/mbedtls" ]; then
         (cd ${BUILDTOP} && rm -rf i386/mbedtls && mkdir -p i386/mbedtls && cd i386/mbedtls && cmake -DCMAKE_C_FLAGS:STRING=-m32 -DCMAKE_INSTALL_PREFIX=$BUILDTOP/i386 ../../mbedtls && make CFLAGS='-m32 --coverage -g3 -O0' && make install)
     fi
 
@@ -31,13 +47,13 @@ then
     if [ ! -d ${BUILDTOP}/libpcap ]; then (cd ${BUILDTOP} && git clone -b libpcap-1.8.1 https://github.com/the-tcpdump-group/libpcap.git ); fi
     if [ ! -d ${BUILDTOP}/tcpdump ]; then (cd ${BUILDTOP} && git clone -b tcpdump-4.8.1 https://github.com/the-tcpdump-group/tcpdump.git ); fi
 
-    if [ ! -d "${BUILDTOP}/x86_64/tcpdump-4.8.1/." ]; then
+    if $build64 && [ ! -d "${BUILDTOP}/x86_64/tcpdump-4.8.1/." ]; then
         (cd ${BUILDTOP} && mkdir -p x86_64/libpcap-1.8.1 && cd x86_64/libpcap-1.8.1 && CFLAGS=-m64 ../../libpcap/configure --prefix=$HOME/stuff --target=x86_64-linux-gnu && make LDFLAGS=-m64 CFLAGS="-m64 -fPIC")
         (cd ${BUILDTOP} && ln -f -s x86_64/libpcap-1.8.1 libpcap && mkdir -p x86_64/tcpdump-4.8.1 && cd x86_64/tcpdump-4.8.1 && CFLAGS=-m64 ../../tcpdump/configure --prefix=$HOME/stuff --target=x86_64-linux-gnu && make LDFLAGS=-m64 CFLAGS="-m64 -fPIC")
         (cd ${BUILDTOP}/x86_64 && ln -s -f tcpdump-4.8.1 tcpdump )
     fi
 
-    if [ ! -d "${BUILDTOP}/i386/tcpdump-4.8.1/." ]; then
+    if $build32 && [ ! -d "${BUILDTOP}/i386/tcpdump-4.8.1/." ]; then
         (cd ${BUILDTOP} && mkdir -p i386/libpcap-1.8.1 && cd i386/libpcap-1.8.1 && CFLAGS=-m32 ../../libpcap/configure --prefix=$HOME/stuff --target=i686-pc-linux-gnu && make LDFLAGS=-m32 CFLAGS="-m32 -fPIC")
         (cd ${BUILDTOP} && ln -f -s i386/libpcap-1.8.1 libpcap && mkdir -p i386/tcpdump-4.8.1 && cd i386/tcpdump-4.8.1 && CFLAGS=-m32 ../../tcpdump/configure --prefix=$HOME/stuff --target=i686-pc-linux-gnu && make LDFLAGS=-m32 CFLAGS="-m32 -fPIC")
         (cd ${BUILDTOP}/i386 && ln -s -f tcpdump-4.8.1 tcpdump )
@@ -62,13 +78,11 @@ then
 
     (cd ${BUILDTOP} && mkdir -p ${HOST}/libcbor && cd ${HOST}/libcbor && cmake ../../libcbor -DCMAKE_INSTALL_PREFIX:PATH=${BUILDTOP} && make && make install)
 
-
-
-    if [ ! -d "${BUILDTOP}/x86_64/libcbor" ]; then
+    if $build64 && [ ! -d "${BUILDTOP}/x86_64/libcbor" ]; then
         (cd ${BUILDTOP} && mkdir -p x86_64/libcbor && cd x86_64/libcbor && cmake ../../libcbor -DCMAKE_C_FLAGS:STRING="-m64 -fPIC" -DCMAKE_INSTALL_PREFIX:PATH=${BUILDTOP}/x86_64 && make && make install)
     fi
 
-    if [ ! -d "${BUILDTOP}/i386/libcbor" ]; then
+    if $build32 && [ ! -d "${BUILDTOP}/i386/libcbor" ]; then
         (cd ${BUILDTOP} && mkdir -p i386/libcbor && cd i386/libcbor && cmake ../../libcbor -DCMAKE_C_FLAGS:STRING="-m32 -fPIC" -DCMAKE_INSTALL_PREFIX:PATH=${BUILDTOP}/i386 && make && make install)
     fi
 fi
